HM-19 Create properties level permit all request matchers.

+++ src/main/java/com/laarc/hoamanagerserver/shared/model/UserRole.java
@@ -16,7 +16,7 @@ public class UserRole {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userRoleId;
+    private Short userRoleId;
 
     @Enumerated(EnumType.STRING)
     private UserRoleName roleName;
